feat(nvim)!: drop support for nvim-lspconfig in favor of native configurations
