+++ config/nvim/.luarc.json
@@ -1,0 +1,3 @@
+{
+
+}
