+++ config/nvim/lsp/luals.lua
@@ -1,0 +1,25 @@
+return {
+  cmd = { "lua-language-server" },
+  filetypes = { "lua" },
+  root_markers = { ".luarc.json", ".luarc.jsonc" },
+  settings = {
+    Lua = {
+      runtime = { version = "LuaJIT" },
+      diagnostics = {
+        globals = { "vim" },
+        disable = {
+          "missing-fields",
+          "incomplete-signature-doc",
+        },
+        unusedLocalExclude = { "_*" },
+      },
+      workspace = {
+        checkThirdParty = false,
+        library = {
+          vim.env.VIMRUNTIME,
+        },
+      },
+      telemetry = { enable = false },
+    },
+  },
+}
