+++ config/nvim/lsp/marksman.lua
@@ -1,0 +1,6 @@
+return {
+  cmd = { "marksman" },
+  filetypes = { "markdown" },
+  root_markers = { ".marksman.toml" },
+  settings = {},
+}
