+++ config/nvim/lua/config.lua
@@ -1,6 +1,5 @@
 require("config.autocmd")
 require("config.keymaps")
-require("config.lsp")
 require("config.options")
 require("config.statusline")
 require("config.usercmds")
