--- config/nvim/lua/config/lsp.lua
@@ -1,45 +1,0 @@
--- constants
-local LSP_SIGNS = {
-  { name = "DiagnosticSignError", text = "" },
-  { name = "DiagnosticSignWarn", text = "" },
-  { name = "DiagnosticSignHint", text = "" },
-  { name = "DiagnosticSignInfo", text = "" },
-}
-
--- configuration for handlers
-vim.lsp.handlers["textDocument/publishDiagnostics"] = vim.lsp.with(vim.lsp.diagnostic.on_publish_diagnostics, {
-  underline = false,
-  virtual_text = false,
-  signs = true,
-  update_in_insert = false,
-})
-
-vim.lsp.handlers["textDocument/hover"] = vim.lsp.with(vim.lsp.handlers.hover, {
-  border = "rounded",
-})
-
-vim.lsp.handlers["textDocument/signatureHelp"] = vim.lsp.with(vim.lsp.handlers.signature_help, {
-  border = "rounded",
-})
-
--- other configurations
-for _, sign in ipairs(LSP_SIGNS) do
-  vim.fn.sign_define(sign.name, { texthl = sign.name, text = sign.text })
-end
-
-vim.diagnostic.config({
-  virtual_text = false,
-  signs = { active = LSP_SIGNS },
-  update_in_insert = true,
-  underline = false,
-  severity_sort = true,
-  float = {
-    focusable = false,
-    style = "minimal",
-    border = "rounded",
-    source = "if_many",
-    header = "",
-    prefix = "",
-    suffix = "",
-  },
-})
