+++ config/nvim/lua/config/options.lua
@@ -25,6 +25,8 @@ vim.opt.clipboard = "unnamedplus"
 vim.opt.completeopt = { "menu", "menuone", "noinsert", "noselect" }
 vim.opt.laststatus = 3
 vim.opt.mouse = ""
+vim.opt.winborder = "rounded"
+vim.opt.mopt = "hit-enter,history:500,wait:500"
 -- backup related
 vim.opt.backup = false
 vim.opt.writebackup = false
